requirements.txt
pygame>=2.5.0
pydantic>=2.0.0
pyyaml>=6.0
openai>=1.0.0
python-dotenv>=1.0.0
requests>=2.28.0
fastapi>=0.110.0
uvicorn>=0.23.0

